requirements.txt
docopt==0.6.2
numpy==1.14.0
Pillow==5.0.0
pydicom==0.9.9
